--- a/lib/parseURL.js
+++ b/lib/parseURL.js
@@ -17,18 +17,18 @@ module.exports = function parseURL (url, query) {
     : new URL(url, BASE_URL)
 
   const merged = Object.assign({}, url.query, query)
-  Object.keys(merged).forEach(key => {
+  for (const key in merged) {
     const value = merged[key]
 
     if (Array.isArray(value)) {
       result.searchParams.delete(key)
-      value.forEach(param => {
+      for (const param of value) {
         result.searchParams.append(key, param)
-      })
+      }
     } else {
       result.searchParams.set(key, value)
     }
-  })
+  }
 
   return result
 }