--- a/lib/request.js
+++ b/lib/request.js
@@ -64,11 +64,12 @@ function Request (options) {
 
   this.headers = {}
   const headers = options.headers || {}
-  Object.keys(headers).forEach((field) => {
+
+  for (const field in headers) {
     const value = headers[field]
     assert(value !== undefined, 'invalid value "undefined" for header ' + field)
     this.headers[field.toLowerCase()] = '' + value
-  })
+  }
 
   this.headers['user-agent'] = this.headers['user-agent'] || 'lightMyRequest'
   this.headers.host = this.headers.host || options.authority || hostHeaderFromURL(parsedURL)
@@ -94,13 +95,15 @@ function Request (options) {
 
   // we keep both payload and body for compatibility reasons
   let payload = options.payload || options.body || null
-  if (payload && typeof payload !== 'string' && !(typeof payload.resume === 'function') && !Buffer.isBuffer(payload)) {
+  const payloadResume = payload && typeof payload.resume === 'function'
+
+  if (payload && typeof payload !== 'string' && !payloadResume && !Buffer.isBuffer(payload)) {
     payload = JSON.stringify(payload)
     this.headers['content-type'] = this.headers['content-type'] || 'application/json'
   }
 
   // Set the content-length for the corresponding payload if none set
-  if (payload && !(typeof payload.resume === 'function') && !this.headers.hasOwnProperty('content-length')) {
+  if (payload && !payloadResume && !this.headers.hasOwnProperty('content-length')) {
     this.headers['content-length'] = (Buffer.isBuffer(payload) ? payload.length : Buffer.byteLength(payload)).toString()
   }