--- a/package.json
+++ b/package.json
@@ -13,6 +13,7 @@
   "types": "index.d.ts",
   "devDependencies": {
     "@types/node": "^16.0.0",
+    "benchmark": "^2.1.4",
     "end-of-stream": "^1.4.4",
     "form-auto-content": "^2.0.0",
     "form-data": "^4.0.0",