Before changes
```
Request x 143,422 ops/sec ±1.22% (85 runs sampled)
Request With Cookies x 121,947 ops/sec ±0.85% (89 runs sampled)
Request With Cookies n payload x 115,137 ops/sec ±1.23% (89 runs sampled)
ParseUrl x 292,202 ops/sec ±0.92% (94 runs sampled)
ParseUrl and query x 210,294 ops/sec ±1.12% (93 runs sampled)
```

After changes:
```
Request x 190,980 ops/sec ±0.82% (90 runs sampled)
Request With Cookies x 150,045 ops/sec ±0.62% (95 runs sampled)
Request With Cookies n payload x 141,071 ops/sec ±1.04% (94 runs sampled)
ParseUrl x 288,184 ops/sec ±0.56% (94 runs sampled)
ParseUrl and query x 211,901 ops/sec ±1.28% (95 runs sampled)
```

I guess the changes in ParseUrl is not significant to make a drastic change in perf, but it fits the coding style.

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)